refactor: convert test files to JS modules
