+++ vault-plugins/flant_iam/tests/tests/01_tenants.mjs
@@ -1,139 +1,10 @@
-const { expect } = require("chai")
-const {
-    Worder,
-    expectStatus,
-    getClient,
-    rootToken,
-    anotherToken,
-} = require("./lib")
-
-const TEN = "tenant"
-const tenId = (id) => TEN + "/" + id
-
-class CRUD {
-    constructor(client) {
-        this.client = client
-    }
-
-    get(endpoint, opts) {
-        return this.client.get(endpoint, opts)
-    }
-
-    post(endpoint, payload, opts) {
-        return this.client.post(endpoint, payload, opts)
-    }
-
-    delete(endpoint, opts) {
-        return this.client.delete(endpoint, opts)
-    }
-}
-
-class API {
-    constructor(client, endpointBuilder) {
-        this.client = new CRUD(client)
-        this.endpointBuilder = endpointBuilder
-    }
-
-    create({ params = {}, query = {}, payload, opts = {} } = {}) {
-        const endpoint = this.endpointBuilder.create(params, query)
-        return this.client.post(endpoint, payload, {
-            ...expectStatus(201),
-            ...opts,
-        })
-    }
-
-    read({ params = {}, query = {}, opts = {} } = {}) {
-        const endpoint = this.endpointBuilder.read(params, query)
-        return this.client.get(endpoint, {
-            ...expectStatus(200),
-            ...opts,
-        })
-    }
-
-    update({ params = {}, query = {}, payload, opts = {} } = {}) {
-        const endpoint = this.endpointBuilder.update(params, query)
-        return this.client.post(endpoint, payload, {
-            ...expectStatus(200),
-            ...opts,
-        })
-    }
-
-    delete({ params = {}, query = {}, opts = {} } = {}) {
-        const endpoint = this.endpointBuilder.delete(params, query)
-        return this.client.delete(endpoint, {
-            ...expectStatus(204),
-            ...opts,
-        })
-    }
-
-    list({ params = {}, query = {}, opts = {} } = {}) {
-        const endpoint = this.endpointBuilder.list(params, query)
-        return this.client.get(endpoint, {
-            ...expectStatus(200),
-            ...opts,
-        })
-    }
-}
-
-class TenantEndpointBuilder {
-    create(p = {}, q = {}) {
-        return "/tenant"
-    }
-
-    read(p = {}, q = {}) {
-        return `/tenant/${p.tenant}`
-    }
-
-    update(p = {}, q = {}) {
-        return `/tenant/${p.tenant}`
-    }
-
-    delete(p = {}, q = {}) {
-        return `/tenant/${p.tenant}`
-    }
-
-    list(p = {}, q = {}) {
-        return "/tenant?list=true"
-    }
-}
-
-class TenantAPI {
-    constructor(client) {
-        this.api = new API(client, new TenantEndpointBuilder())
-    }
-
-    create(payload, opts) {
-        return this.api.create({ payload, opts })
-    }
-
-    read(id, opts) {
-        return this.api.read({ params: { tenant: id }, opts })
-    }
-
-    update(id, payload, opts) {
-        return this.api.update({ params: { tenant: id }, payload, opts })
-    }
-
-    delete(id, opts) {
-        return this.api.delete({ params: { tenant: id }, opts })
-    }
-
-    list(opts) {
-        return this.api.list({ opts })
-    }
-}
+import { expectStatus, getClient, rootToken } from "./lib/client.mjs"
+import { genTenantPayload, TenantAPI } from "./lib/tenant.mjs"
+import { expect } from "chai"
 
 describe("Tenants", function () {
     const rootClient = getClient(rootToken)
     const root = new TenantAPI(rootClient)
-    const worder = new Worder()
-
-    function genTenantPayload(override = {}) {
-        return {
-            name: worder.gen(),
-            ...override,
-        }
-    }
 
     // afterEach("cleanup", async function () {
     //     const clean = s => root.delete(`tenant/${s}`, expectStatus(204))
